requirements.txt
asgiref==3.5.2
black==22.10.0
click==8.1.3
Django==4.1.1
djangorestframework==3.14.0
gunicorn==20.1.0
mypy-extensions==0.4.3
numpy==1.23.4
opencv-python==4.6.0.66
packaging==21.3
pathspec==0.10.1
Pillow==9.2.0
platformdirs==2.5.2
pyparsing==3.0.9
pytesseract==0.3.10
pytz==2022.2.1
sqlparse==0.4.3
tomli==2.0.1
